requirements.txt
openai
pinecone-client
python-dotenv
pandas
langchain
numpy
textwrap3
streamlit
pdfplumber